fix: 修复PR #12 CI失败的C++测试问题
- AdaptiveSpinTest.SpinPhase: 修复自比较断言 (idle_count < idle_count)
- FastClockTest: 修正Windows平台时间戳期望值 (QPC返回系统启动时间)
- fast_clock.cpp: 使用double中间值避免Windows QPC整数溢出

本地验证: Python 162 tests passed, C++ 329 tests passed

+++ src/tyche/cpp/engine/fast_clock.cpp
@@ -34,8 +34,10 @@ int64_t FastClock::_system_now_ns() noexcept {
     QueryPerformanceFrequency(&freq);
     QueryPerformanceCounter(&count);
     // Convert to nanoseconds: count * 1e9 / freq
+    // Use double intermediate to avoid overflow on high counter values
     return static_cast<int64_t>(
-        (count.QuadPart * 1'000'000'000LL) / freq.QuadPart);
+        static_cast<double>(count.QuadPart) * 1'000'000'000.0 /
+        static_cast<double>(freq.QuadPart));
 #else
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
